Move build spectral library to a utility function
